+++ Reni/CodeVisitor.h
@@ -21,12 +21,12 @@ namespace Reni
         virtual String const NumberOperation(String const& name, Size const&size, int leftDepth, Size const&leftSize, int rightDepth, Size const&rightSize)const;
         virtual String const NumberConversion(Size const& size, int argDepth, Size const& argSize) const;
         virtual String const FunctionArg() const;
-        virtual String const CallGetter(Size const&result, int const index, Size const&args) const;
+        virtual String const CallGetter(Size const&result, size_t const index, Size const&args) const;
         
-        static String const InName(String const&prefix, int index);
-        static String const GetterFunctionDeclaration(int const index);
-        static String const GetterName(int const index){ return "get_" + String::Convert(index); };
-        static String const SetterName(int const index){ return "set_" + String::Convert(index); };
+        static String const InName(String const&prefix, size_t index);
+        static String const GetterFunctionDeclaration(size_t const index);
+        static String const GetterName(size_t const index){ return "get_" + String::Convert(index); };
+        static String const SetterName(size_t const index){ return "set_" + String::Convert(index); };
 
     };
 
@@ -38,12 +38,12 @@ namespace Reni
         String const DumpPrintNumber(Size const size) const override;
         String const FiberConnection(Array<Ref<CodeItem>> const&items, Ref<FiberConnectorItem> const&connector) const override;
         String const NumberOperation(String const& name, Size const&size, int leftDepth, Size const&leftSize, int rightDepth, Size const&rightSize)const override;
-        String const CallGetter(Size const& result, int const index, Size const& args) const override;
+        String const CallGetter(Size const& result, size_t const index, Size const& args) const override;
         String const FunctionArg() const override;
         String const Visit(Ref<CodeItem> target)const;
     public:
         static String const MainVisit(Ref<CodeItem> body);
-        static String const GetterVisit(int index, Ref<CodeItem> body);
+        static String const GetterVisit(size_t index, Ref<CodeItem> body);
     };
 
 }
